pyproject.toml
[tool.poetry]
name = "rcm-os"
version = "0.1.0"
description = "AI-native, human-assisted RCM OS (MVP)"
authors = ["ModernRCM Team"]
readme = "README.md"
packages = [
  { include = "backend" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {version = "^0.30.0", extras = ["standard"]}
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg = {version = "^3.1.19", extras = ["binary"]}
python-dotenv = "^1.0.1"
boto3 = "^1.34.0"
redis = "^5.0.4"
temporalio = "^1.6.0"
pyyaml = "^6.0.1"
opentelemetry-sdk = "^1.25.0"
opentelemetry-instrumentation-fastapi = "^0.46b0"
prometheus-client = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
ruff = "^0.5.0"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
